+++ packages/terra-data-grid/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Changed
+  * Updated prop descriptions for better consistency and accuracy in the `terra-data-grid` component.
+
 ## 0.7.0 - (October 3, 2023)
 
 * Fixed